Add syntax highlighting option to post and page archetype
This option, default = false, enables or disables the loading of the
CSS and Javascript files required for syntax highlighting. This means
that those files are loaded only in the header and footer of the pages
and posts that require it explicitly.

Result: faster loading pages and especially the home page.