Move render blocking CSS to footer according to Google PageSpeed
A Javascript inline script loads the two stylesheets (the minimized
for the website and the font stylesheet) without blocking the rendering.
If the Javascript is disabled, the two stylesheets are loaded in a
render blocking old-school way with the <link> tag inside the <noscript>
one. Still this tag is in the footer of the page to make mobile browsers
render the page before any CSS is loaded (this improves the mobile
PageSpeed score from 75 to 100).