Substitute png logo with inline-svg version
This removes one connection for the page loading, any complication for
retina and non-retina screens and reduces the image size at ~90% and
~36% of the png image size for the non-retina and retina versions
respectively.

The CSS has been slightly modified to apply the same effect to the svg
logo as well.

A fallback png image has to be set up for browsers which don't support
the inline svg.