Serve Highlight.js from same host, unify its CSS into style.css
This makes the website entirely hosted locally, there are no external
resources.

The code highlighting library Highlight.js is served from the same
domain as the website, while its CSS code is incorporated in the
unified style.css file, since minified since it's just 600 B in size.

On the other hand the Javascript file required for highlighting will
still be called only when needed (when posts and pages require it in the
frontmatter) since it's 40 kB in size.