Add custom RSS feed layout with post description instead of content
The only change compared to the embedded RSS feed in Hugo is the line 20
where ".Content" is replaced with ".Description". This shows the excerpt
of the post to be seen instead of the full post text, much like what
happens on the home page.

This reduces the feed file size from 127.7 kB to 9.3 kB uncompressed,
not counting the images included in the content that any browser or
feed reader may load.