Change filename scheme from ISO date to serial number
The change maintains the same ordering of the files but also avoids
data duplication, as the publishing date is available inside the
frontmatter of the post.

Another problem that is solved is a post that is the date mismatch
for posts that are written on a certain date and published on another.
With this solution drafts can also have the number 000.