Replace summary_large Twitter card with summary, added appropriate image
The summary_large_image Twitter card requires a specific image for each
post which I don’t want to use on my website (for now). This leads me
to use a summary Twitter card instead with the logo of the website. The
logo is a JPG image to avoid problems with the transparent background on
the Twitter night mode or different backgrounds.