Fixes in highlighting library calls, whitespace, edit notes
After the enabling in the horizontal scrolling of the code blocks where
no highlighting was required in the first place, the calls to the
highlightin library are removed. The `{{< code >}}` shortcode usage
is also removed. When possible the loading of the highlighting library
in the frontmatter is deactivated.

After some new languages were added to the highlighting library, those
were enabled where needed, such as `apache` for the `.htaccess` logs
instead of `nginx` or `HTTP` instead of `text`.

Some non-breaking whitespace in older posts is removed in order to
make the formatting (italic, bold) work properly.

Unified the formatting of the _Edit_ notes in the footers of the posts
as much as possible in something with the format _Edit YYYY-MM-DD:_